48 hour check works

+++ frontend/src/pages/AllBookings.js
@@ -3,6 +3,7 @@ import React, { useState, useEffect } from 'react';
 const AllBookings =()=>{
 
 const [bookings, setBookings] = useState([]);
+const [hasError, setHasError] = useState(false);
 
     useEffect(() => {
         loadData();
@@ -14,34 +15,50 @@ const [bookings, setBookings] = useState([]);
 
     const loadData = async () => {
         const response = await fetch ("http://localhost:8080/booking/viewAll")
+
             .then(response => response.json())
             .then(data => setBookings(data))
+            .catch(error => setHasError(true))
 
-            //TODO error handling and sessions check
+            //TODO sessions check
 
     };
 
     const deleteBooking = (entry) =>{
-        //TODO 48h check
+        if(dateCheck(entry.date)){
+            alert("Du kan inte avboka en städning mindre än 48 timmar innan bokat tid");
+        }
+        else{
+             bookings.map(booking => {
+                if(booking.id === entry.id){
+                    //console.log(entry.id);
+                    fetch("http://localhost:8080/booking/delete/" + entry.id, {
+                        method: "delete"
+                    })
+                    setBookings(bookings.filter(item => item.id !== entry.id));
+                }})
+        }
 
-        bookings.map(booking => {
-            if(booking.id === entry.id){
-                console.log(entry.id);
-                fetch("http://localhost:8080/booking/delete/" + entry.id, {
-                    method: "delete"
-                })
-                setBookings(bookings.filter(item => item.id !== entry.id));
-            }
-        })
     }
 
-
-    /*
     const dateCheck = (date)=>{
         let today = new Date();
-        console.log(date);
+        let bookingDate = new Date(date);
+        let diffInHours = Math.abs(today-bookingDate)/ (60*60*1000);
 
-    }*/
+       //check if difference is more than 48H
+        if(diffInHours <= 48){
+            return true; //cannot cancel
+        }else{
+            return false //ok to cancel
+        }
+    }
+
+    const convertDateAndTime= (date)=>{
+        let convertToLocale = new Date(date).toLocaleDateString('sv-SE', {year:"numeric", month:"numeric", day: "numeric", hour: "numeric", minute: "2-digit"});
+        return convertToLocale;
+    }
+
 
     return (
         <div className="flex flex-col">
@@ -52,21 +69,25 @@ const [bookings, setBookings] = useState([]);
                             <thead className="bg-gray-50">
                                 <tr>
                                   <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                    Name</th>
+                                    Namn</th>
                                   <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                    Date</th>
+                                    Datum</th>
                                   <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                                    Cleaning Service</th>
+                                    Tid</th>
+                                  <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                                    Städtjänst</th>
                                   <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
                                     Adress</th>
                                   <th></th>
                                 </tr>
                             </thead>
                             <tbody className="bg-white divide-y divide-gray-200">
-                                {bookings.map((entry) =>
+
+                                {hasError? <div>Nåt gick fel</div> : (bookings.map((entry) =>
                                 <tr key={entry.id}>
                                     <td className="px-6 py-4 text-left whitespace-nowrap text-sm text-gray-900">{entry.name}</td>
-                                    <td className="px-6 py-4 text-left whitespace-nowrap text-sm text-gray-900">{entry.date}</td>
+                                    <td className="px-6 py-4 text-left whitespace-nowrap text-sm text-gray-900">{convertDateAndTime(entry.date).split(" ", 1)}</td>
+                                    <td className="px-6 py-4 text-left whitespace-nowrap text-sm text-gray-900">{convertDateAndTime(entry.date).split(" ").slice(1)}</td>
                                     <td className="px-6 py-4 text-left whitespace-nowrap text-sm text-gray-900">{entry.cleaningPackage}</td>
                                     <td className="px-6 py-4 text-left whitespace-nowrap text-sm text-gray-900">{entry.address}</td>
 
@@ -76,7 +97,7 @@ const [bookings, setBookings] = useState([]);
 
                                     </td>
                                 </tr>
-                            )}
+                            ))}
                             </tbody>
                         </table>
                     </div>
